+++ core/datastore/src/main/kotlin/io/github/vedicmitra/core/datastore/di/DataStoreModule.kt
@@ -22,10 +22,12 @@ import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.android.qualifiers.ApplicationContext
 import dagger.hilt.components.SingletonComponent
+import io.github.vedicmitra.core.datastore.DefaultJapaRepository
 import io.github.vedicmitra.core.datastore.DefaultLocationRepository
 import io.github.vedicmitra.core.datastore.DefaultProfileRepository
 import io.github.vedicmitra.core.datastore.DefaultReminderRepository
 import io.github.vedicmitra.core.datastore.DefaultUserPreferencesRepository
+import io.github.vedicmitra.core.datastore.JapaRepository
 import io.github.vedicmitra.core.datastore.LocationRepository
 import io.github.vedicmitra.core.datastore.ProfileRepository
 import io.github.vedicmitra.core.datastore.ReminderRepository
@@ -55,6 +57,9 @@ internal abstract class DataStoreModule {
 
     @Binds
     abstract fun bindProfileRepository(impl: DefaultProfileRepository): ProfileRepository
+
+    @Binds
+    abstract fun bindJapaRepository(impl: DefaultJapaRepository): JapaRepository
 
     companion object {
         @Provides